Add vplot external utility for plot.v example
- Implement vplot: 3D surface plotter and OFF model viewer for X11
- Add TTY class support for macOS using forkpty()
- Fix 64-bit pointer handling in YYSTYPE union (int -> long)
- Add defensive pointer validation in saveString()
- Include X-29 and F-14 TomCat 3D models in OFF format
- X-29 model: Copyright (c) Evans & Sutherland
- Remove debug output from cl_TTY.c and glib_x.c

+++ src/viola/y.tab.h
@@ -39,7 +39,7 @@
 
 typedef union {
     struct AST* p; /* an AST node	*/
-    int i;
+    long i;
     float f;
     char c;
     char* s;
